move old code to typescript app
