+++ components/gcp-click-to-deploy/src/logo.jpg~HEAD
