+++ components/gcp-click-to-deploy/tslint.json
@@ -11,7 +11,6 @@
     ]
   },
   "rules": {
-    "interface-name": false,
     "max-classes-per-file": false
   }
 }
