+++ components/gcp-click-to-deploy/tslint.json~HEAD
@@ -11,6 +11,7 @@
     ]
   },
   "rules": {
+    "interface-name": false,
     "max-classes-per-file": false
   }
 }
