--- components/gcp-deployer-ts/module_types.d.ts
@@ -1,5 +1,0 @@
-declare module '*.jinja'
-declare module '*.jpg'
-declare module '*.png'
-declare module '*.svg'
-declare module '*.yaml'
