--- components/gcp-deployer-ts/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "gcp-deployer",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "js-yaml": "^3.12.0",
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0",
-    "react-scripts-ts": "2.16.0"
-  },
-  "scripts": {
-    "start": "react-scripts-ts start",
-    "build": "react-scripts-ts build",
-    "test": "react-scripts-ts test --env=jsdom",
-    "eject": "react-scripts-ts eject"
-  },
-  "devDependencies": {
-    "@types/gapi": "0.0.35",
-    "@types/gapi.auth2": "0.0.47",
-    "@types/jest": "^23.0.0",
-    "@types/js-yaml": "^3.11.1",
-    "@types/node": "^10.3.2",
-    "@types/react": "^16.3.17",
-    "@types/react-dom": "^16.0.6",
-    "typescript": "^2.9.1"
-  }
-}
