--- components/gcp-deployer-ts/src/App.css
@@ -1,27 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 80px;
-}
-
-.App-header {
-  background-color: #222;
-  color: white;
-  height: 150px;
-  padding: 20px;
-}
-
-.App-title {
-  font-size: 1.5em;
-}
-
-.App-intro {
-  font-size: large;
-}
-
-.NameForm {  
-  height: 304px;
-  width: 704px;
-}
