--- components/gcp-deployer-ts/src/configs/cluster.jinja.schema
@@ -1,42 +1,0 @@
-# Copyright 2016 Google Inc. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-info:
-  title: GKE cluster
-  author: Google, Inc.
-  description: |
-    Creates a GKE cluster and associated type for use in DM. The type can be
-    used in other DM configurations in the following manner:
-
-    "type: <cluster-type>:/api/v1/namespaces/{namespace}/services"
-
-required:
-- zone
-
-properties:
-  zone:
-    type: string
-    description: Zone in which the cluster should run.
-  initialNodeCount:
-    type: integer
-    description: Initial number of nodes desired in the cluster.
-    default: 4
-
-outputs:
-  clusterType:
-    description: The name of the type provider which can create resources from the Kubernetes v1 API in your cluster.
-    type: string
-  clusterType-v1beta1-extensions:
-    description: The name of the type provider which can create resources from the Kubernetes v1beta1-extensions API in your cluster.
-    type: string
