--- components/gcp-deployer-ts/src/index.tsx
@@ -1,9 +1,0 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-import App from './App';
-import './index.css';
-
-ReactDOM.render(
-  <App />,
-  document.getElementById('root') as HTMLElement
-);
