--- components/gcp-deployer-ts/tsconfig.prod.json
@@ -1,3 +1,0 @@
-{
-  "extends": "./tsconfig.json"
-}
